MidiGenerator - create add and remove neuron function on MidiGenerator
This also starts the work to make the updating of the brain thead safe.

Adding a neuron to the brain requires the updating of state in multiple
places. For this reason, in order to add a new neuron we need to create
a copy of the brain and the swap out a pointer when it is ready to be
used.